+++ src/shared/codelens/typescriptCodeLensProvider.ts
@@ -30,7 +30,8 @@ import {
 } from './localLambdaRunner'
 
 const unsupportedNodeJsRuntimes: Set<string> = new Set<string>([
-    'nodejs4.3'
+    'nodejs4.3',
+    'nodejs6.10',
 ])
 
 async function getSamProjectDirPathForFile(filepath: string): Promise<string> {
@@ -70,8 +71,6 @@ export function initialize({
             debugPort = await getDebugPort()
         }
 
-        const protocol = params.runtime === 'nodejs6.10' ? 'legacy' : 'inspector'
-
         const debugConfig: NodejsDebugConfiguration = {
             type: 'node',
             request: 'attach',
@@ -81,7 +80,7 @@ export function initialize({
             port: debugPort!,
             localRoot: samProjectCodeRoot,
             remoteRoot: '/var/task',
-            protocol,
+            protocol: 'inspector',
             skipFiles: [
                 '/var/runtime/node_modules/**/*.js',
                 '<node_internals>/**/*.js'
@@ -114,11 +113,11 @@ export function initialize({
             })
             const runtime = CloudFormation.getRuntime(resource)
 
-            if (params.isDebug && unsupportedNodeJsRuntimes.has(runtime)) {
+            if (unsupportedNodeJsRuntimes.has(runtime)) {
                 vscode.window.showErrorMessage(
                     localize(
-                        'AWS.lambda.debug.runtime.unsupported',
-                        'Debug support for {0} is currently not supported',
+                        'AWS.samcli.local.invoke.runtime.unsupported',
+                        'Local invoke with {0} is not supported',
                         runtime
                     )
                 )